Properly refactor SContext/IntFieldContext.
These are now SScope and IntFieldScope, keeping with specs2 terminology, and
properly wrap each of the specs that use them individually to provide
isolation (though we could use `isolated` for that as well). All tests pass
now, boom!
