+++ src/test/scala/com/hacklanta/formality/FieldSpec.scala
@@ -3,10 +3,8 @@ package formality
 
 import scala.xml._
 
-import org.specs2.execute.AsResult
 import org.specs2.matcher.XmlMatchers._
 import org.specs2.mutable._
-import org.specs2.specification.AroundEach
 
 import net.liftweb.common._
 import net.liftweb.http._
@@ -14,22 +12,11 @@ import net.liftweb.util._
 
 import Formality._
 
-trait SessionContext {
-  val session = new LiftSession("", StringHelpers.randomString(20), Empty)
-}
-trait SContext extends AroundEach with SessionContext {
-  def around[T : AsResult](t: =>T) = {
-    S.initIfUninitted(session) {
-      AsResult(t)  // execute t inside a http session
-    }
-  }
-}
-
-class FieldSpec extends Specification with SContext {
+class FieldSpec extends Specification {
   val templateElement = <div class="boomdayada boomdayadan" data-test-attribute="bam">Here's a test!</div>
 
   "Simple fields with no initial value" should {
-    "only bind the name attribute" in {
+    "only bind the name attribute" in new SScope {
       val formField = field[String](".boomdayada")
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -45,7 +32,7 @@ class FieldSpec extends Specification with SContext {
   }
   
   "Simple fields with an initial value" should {
-    "only bind the name and value attributes" in {
+    "only bind the name and value attributes" in new SScope {
       val formField = field[String](".boomdayada", "Dat value")
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -61,7 +48,7 @@ class FieldSpec extends Specification with SContext {
   }
 
   "Regular file upload fields" should {
-    "only bind the name and type attributes" in {
+    "only bind the name and type attributes" in new SScope {
       val formField = fileUploadField(".boomdayada")
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -77,7 +64,7 @@ class FieldSpec extends Specification with SContext {
   }
 
   "Typed file upload fields" should {
-    "only bind the name and type attributes" in {
+    "only bind the name and type attributes" in new SScope {
       implicit def fileToObject(fph: FileParamHolder) = Full("boom")
 
       val formField = typedFileUploadField[String](".boomdayada")
@@ -101,7 +88,7 @@ class FieldSpec extends Specification with SContext {
       SHtml.SelectableOption(new Exception("slabai"), "slabai")
     )
 
-    "replace the element wholesale with a select element" in {
+    "replace the element wholesale with a select element" in new SScope {
       val formField = selectField[Exception](".boomdayada", objects)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -113,7 +100,7 @@ class FieldSpec extends Specification with SContext {
         "name" -> ".*"
       )
     }
-    "carry any SelectableOption attributes into the resulting options" in {
+    "carry any SelectableOption attributes into the resulting options" in new SScope {
       val objects = List(
         SHtml.SelectableOption(new Exception("ohai"), "ohai", ("test" -> "bam")),
         SHtml.SelectableOption(new Exception("obai"), "obai", ("other-test" -> "bam")),
@@ -137,7 +124,7 @@ class FieldSpec extends Specification with SContext {
         "still-other-test" -> "bam"
       )
     }
-    "mark as selected the default object" in {
+    "mark as selected the default object" in new SScope {
       val default = new Exception("ohai")
       val objects = List(
         SHtml.SelectableOption(default, "ohai"),
@@ -172,7 +159,7 @@ class FieldSpec extends Specification with SContext {
         </li>
       </ul>
 
-    "only bind to radio buttons and labels in the markup" in {
+    "only bind to radio buttons and labels in the markup" in new SScope {
       val formField = selectField[Exception]("li", objects, asRadioButtons = true)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -191,7 +178,7 @@ class FieldSpec extends Specification with SContext {
       // They should all have different values.
       (Set[String]() ++ inputs.map(_ \ "@value").collect { case Group(Seq(Text(value))) => value }).size must_== 3
     }
-    "carry any SelectableOption attributes into the resulting radio buttons" in {
+    "carry any SelectableOption attributes into the resulting radio buttons" in new SScope {
       val objects = List(
         SHtml.SelectableOption(new Exception("ohai"), "ohai", ("test" -> "bam")),
         SHtml.SelectableOption(new Exception("obai"), "obai", ("other-test" -> "bam")),
@@ -215,7 +202,7 @@ class FieldSpec extends Specification with SContext {
         "still-other-test" -> "bam"
       )
     }
-    "set the associated label's for attribute when an option has the id attribute set" in {
+    "set the associated label's for attribute when an option has the id attribute set" in new SScope {
       val objects = List(
         SHtml.SelectableOption(new Exception("ohai"), "ohai", ("id" -> "bam")),
         SHtml.SelectableOption(new Exception("obai"), "obai", ("other-test" -> "bam")),
@@ -235,7 +222,7 @@ class FieldSpec extends Specification with SContext {
         "for" -> "boom"
       ) \> "slabai"
     }
-    "mark as selected the default object" in {
+    "mark as selected the default object" in new SScope {
       val default = new Exception("ohai")
       val objects = List(
         SHtml.SelectableOption(default, "ohai"),
@@ -266,7 +253,7 @@ class FieldSpec extends Specification with SContext {
       (new Exception("slabai"), "slabai")
     )
 
-    "replace the  element wholesale with a select element" in {
+    "replace the  element wholesale with a select element" in new SScope {
       val formField = selectField[Exception](".boomdayada", objects)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -278,7 +265,7 @@ class FieldSpec extends Specification with SContext {
         "name" -> ".*"
       )
     }
-    "mark as selected the default object" in {
+    "mark as selected the default object" in new SScope {
       val default = new Exception("ohai")
       val objects = List(
         (default, "ohai"),
@@ -313,7 +300,7 @@ class FieldSpec extends Specification with SContext {
         </li>
       </ul>
 
-    "only bind to radio buttons and labels in the markup" in {
+    "only bind to radio buttons and labels in the markup" in new SScope {
       val formField = selectField[Exception]("li", objects, asRadioButtons = true)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -332,7 +319,7 @@ class FieldSpec extends Specification with SContext {
       // They should all have different values.
       (Set[String]() ++ inputs.map(_ \ "@value").collect { case Group(Seq(Text(value))) => value }).size must_== 3
     }
-    "mark as selected the default object" in {
+    "mark as selected the default object" in new SScope {
       val default = new Exception("ohai")
       val objects = List(
         (default, "ohai"),
@@ -363,7 +350,7 @@ class FieldSpec extends Specification with SContext {
       "slabai"
     )
 
-    "replace the  element wholesale with a select element" in {
+    "replace the  element wholesale with a select element" in new SScope {
       val formField = selectField[String](".boomdayada", objects)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -375,7 +362,7 @@ class FieldSpec extends Specification with SContext {
         "name" -> ".*"
       )
     }
-    "mark as selected the default object" in {
+    "mark as selected the default object" in new SScope {
       val formField = selectField[String](".boomdayada", objects, Full("ohai"))
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -403,7 +390,7 @@ class FieldSpec extends Specification with SContext {
         </li>
       </ul>
 
-    "only bind to radio buttons and labels in the markup" in {
+    "only bind to radio buttons and labels in the markup" in new SScope {
       val formField = selectField[String]("li", objects, asRadioButtons = true)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -422,7 +409,7 @@ class FieldSpec extends Specification with SContext {
       // They should all have different values.
       (Set[String]() ++ inputs.map(_ \ "@value").collect { case Group(Seq(Text(value))) => value }).size must_== 3
     }
-    "mark as selected the default object" in {
+    "mark as selected the default object" in new SScope {
       val formField = selectField[String]("li", objects, Full("ohai"), asRadioButtons = true)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -447,7 +434,7 @@ class FieldSpec extends Specification with SContext {
       SHtml.SelectableOption(new Exception("slabai"), "slabai")
     )
 
-    "replace the element wholesale with a select element" in {
+    "replace the element wholesale with a select element" in new SScope {
       val formField = multiSelectField[Exception](".boomdayada", objects)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -460,7 +447,7 @@ class FieldSpec extends Specification with SContext {
         "name" -> ".*"
       )
     }
-    "carry any SelectableOption attributes into the resulting options" in {
+    "carry any SelectableOption attributes into the resulting options" in new SScope {
       val objects = List(
         SHtml.SelectableOption(new Exception("ohai"), "ohai", ("test" -> "bam")),
         SHtml.SelectableOption(new Exception("obai"), "obai", ("other-test" -> "bam")),
@@ -484,7 +471,7 @@ class FieldSpec extends Specification with SContext {
         "still-other-test" -> "bam"
       )
     }
-    "mark as selected the default objects" in {
+    "mark as selected the default objects" in new SScope {
       val defaults = List(new Exception("ohai"), new Exception("slabai"))
       val objects = List(
         SHtml.SelectableOption(defaults(0), "ohai"),
@@ -523,7 +510,7 @@ class FieldSpec extends Specification with SContext {
         </li>
       </ul>
 
-    "only bind to checkboxes and labels in the markup" in {
+    "only bind to checkboxes and labels in the markup" in new SScope {
       val formField = multiSelectField[Exception]("li", objects, asCheckboxes = true)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -542,7 +529,7 @@ class FieldSpec extends Specification with SContext {
       // They should all have different values.
       (Set[String]() ++ inputs.map(_ \ "@value").collect { case Group(Seq(Text(value))) => value }).size must_== 3
     }
-    "carry any SelectableOption attributes into the resulting checkboxes" in {
+    "carry any SelectableOption attributes into the resulting checkboxes" in new SScope {
       val objects = List(
         SHtml.SelectableOption(new Exception("ohai"), "ohai", ("test" -> "bam")),
         SHtml.SelectableOption(new Exception("obai"), "obai", ("other-test" -> "bam")),
@@ -566,7 +553,7 @@ class FieldSpec extends Specification with SContext {
         "still-other-test" -> "bam"
       )
     }
-    "set the associated label's for attribute when an option has the id attribute set" in {
+    "set the associated label's for attribute when an option has the id attribute set" in new SScope {
       val objects = List(
         SHtml.SelectableOption(new Exception("ohai"), "ohai", ("id" -> "bam")),
         SHtml.SelectableOption(new Exception("obai"), "obai", ("other-test" -> "bam")),
@@ -586,7 +573,7 @@ class FieldSpec extends Specification with SContext {
         "for" -> "boom"
       ) \> "slabai"
     }
-    "mark as selected the default objects" in {
+    "mark as selected the default objects" in new SScope {
       val defaults = List(new Exception("ohai"), new Exception("slabai"))
       val objects = List(
         SHtml.SelectableOption(defaults(0), "ohai"),
@@ -623,7 +610,7 @@ class FieldSpec extends Specification with SContext {
       (new Exception("slabai"), "slabai")
     )
 
-    "replace the  element wholesale with a select element" in {
+    "replace the  element wholesale with a select element" in new SScope {
       val formField = multiSelectField[Exception](".boomdayada", objects)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -635,7 +622,7 @@ class FieldSpec extends Specification with SContext {
         "name" -> ".*"
       )
     }
-    "mark as selected the default objects" in {
+    "mark as selected the default objects" in new SScope {
       val defaults = List(new Exception("ohai"), new Exception("slabai"))
       val objects = List(
         (defaults(0), "ohai"),
@@ -674,7 +661,7 @@ class FieldSpec extends Specification with SContext {
         </li>
       </ul>
 
-    "only bind to checkboxes and labels in the markup" in {
+    "only bind to checkboxes and labels in the markup" in new SScope {
       val formField = multiSelectField[Exception]("li", objects, asCheckboxes = true)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -693,7 +680,7 @@ class FieldSpec extends Specification with SContext {
       // They should all have different values.
       (Set[String]() ++ inputs.map(_ \ "@value").collect { case Group(Seq(Text(value))) => value }).size must_== 3
     }
-    "mark as selected the default objects" in {
+    "mark as selected the default objects" in new SScope {
       val defaults = List(new Exception("ohai"), new Exception("slabai"))
       val objects = List(
         (defaults(0), "ohai"),
@@ -730,7 +717,7 @@ class FieldSpec extends Specification with SContext {
       "slabai"
     )
 
-    "replace the  element wholesale with a select element" in {
+    "replace the  element wholesale with a select element" in new SScope {
       val formField = multiSelectField[String](".boomdayada", objects)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -742,7 +729,7 @@ class FieldSpec extends Specification with SContext {
         "name" -> ".*"
       )
     }
-    "mark as selected the default objects" in {
+    "mark as selected the default objects" in new SScope {
       val formField = multiSelectField[String](".boomdayada", objects, List("ohai", "slabai"))
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -774,7 +761,7 @@ class FieldSpec extends Specification with SContext {
         </li>
       </ul>
 
-    "only bind to radio buttons and labels in the markup" in {
+    "only bind to radio buttons and labels in the markup" in new SScope {
       val formField = multiSelectField[String]("li", objects, asCheckboxes = true)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -793,7 +780,7 @@ class FieldSpec extends Specification with SContext {
       // They should all have different values.
       (Set[String]() ++ inputs.map(_ \ "@value").collect { case Group(Seq(Text(value))) => value }).size must_== 3
     }
-    "mark as selected the default objects" in {
+    "mark as selected the default objects" in new SScope {
       val formField = multiSelectField[String]("li", objects, List("ohai", "slabai"), asCheckboxes = true)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
@@ -818,7 +805,7 @@ class FieldSpec extends Specification with SContext {
   }
 
   "Checkbox fields with Boolean values" should {
-    "replace the element with a checkbox-hidden input pair" in {
+    "replace the element with a checkbox-hidden input pair" in new SScope {
       val formField = checkboxField(".boomdayada", false)
 
       val resultingMarkup = <test-parent>{formField.binder(templateElement)}</test-parent>
